SPL: Fix build of CONFIG_SPL_NET_SUPPORT
With CONFIG_SPL_NET_SUPPORT set we bring in most of the environment
related code.  This in turn means that while we discard the callback
saftey checks in the environment, we had still needed their
__start/__end linker symbols.  In most cases this had been working
because we generated u-boot.lst for the main U-Boot build prior to
creating u-boot-spl.lds.  With a sufficiently large machine this is not
the case and exposed this latent bug which is that as of f8cfcf1 there
are no callback linker entries so not __start/__end symbol was
generated.

As the environment is not user modifiable in this particular run-time
(for any variable that had a callback associated with it) we simply
provide an empty env_callback_init in SPL.

Cc: Joe Hershberger <joe.hershberger@ni.com>
Cc: Scott Wood <scottwood@freescale.com>
Tested-by: Tom Rini <trini@ti.com> (Ran with am335x_evm_usbspl build)
Signed-off-by: Tom Rini <trini@ti.com>
Acked-by: Joe Hershberger <joe.hershberger@ni.com>