am335x_evm: Add support for 720MHz/800MHz/1GHz
The GP EVM may contain PG1.0, PG2.0 or PG2.1 silicon.  The EVM SK may
contain PG1.0 or PG2.1 silicon.  In the case of PG2.1 and later we have
the ability to determine at runtime the maximum frequency supported and
switch to it.  In the case of PG1.0 and PG2.0 we must simply assume a
maximum.

Signed-off-by: Tom Rini <trini@ti.com>