am33xx:ddr: Boost priority of older commands in EMIF
Sometimes it is observed that when MPU is performing EMIF
accesses largely, the older commands in EMIF never get
a chance to execute. This results in LCD underflow and sync loss
if LCD DMA parellely pulling data from memory.

This issue is fixed by giving a priority boost to old commands
in EMIF command FIFO.

Change-Id: Icad5811dfbab5a57f2dd1ef0bf9e7de3b219827c
Signed-off-by: Arun Joseph <arunjoseph@ti.com>