Default to fastboot if no boot image and add boot reason flags
To better support Android devices, change the auto boot sequence
as follows:

If CONFIG_CMD_FASTBOOT is defined, the boot sequence is determined by
a "boot reason" flag in the top part of SRAM. If the flag is not set
(e.g. because this is a cold boot) or it is set to REBOOT_FLAG_NORMAL
the boot command is "booti mmc1". If the flag is REBOOT_FLAG_RECOVERY
the command is "booti mmc1 recovery". If it is REBOOT_FLAG_FASTBOOT
the command is just "fastboot".

To make use of this you need a Linux kernel which will set the
reboot flag.